requirements.txt
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
emoji==0.5.4
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.3.0
idna==2.9
passlib==1.7.2
pprint==0.1
praw==6.5.1
prawcore==1.0.1
progressbar2==3.50.0
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-utils==2.4.0
requests==2.23.0
six==1.14.0
update-checker==0.16
urllib3==1.25.8
websocket-client==0.57.0
